Beta version of the Satochip second factor authentication:
- basic gui
- basic parse and display of tx
- support scanning of pairing qr code
- support label description for each pairing
- support multiple 2FA (with multiple satochip)
- logs of tx
