+++ make_packages
@@ -1,0 +1,3 @@
+#!/bin/bash
+#Install pure python modules in the directory
+python3 -m pip install -r ./requirements.txt -t .
