+++ requirements.txt
@@ -1,0 +1,6 @@
+certifi
+chardet
+cryptos
+idna
+requests
+urllib3
